First commit
Decided to move this to a separate repository and prepare/test it for
external use. It was lacking in flexibility which I realised when a
friend wanted to use it and it didn't support some simple things like
optional parameters and comments.

Will be updating this after more testing, feel free to submit your own
map coding styles that I can add support for.