Updated parsing code to be more flexible
Now the code splits each line into the function name and the arguments
using sscanf (and removes the closing bracket, ';' character and
optional comment) then checks what the function name is and does further
sscanf queries on the arguments depending on the function name. This
cuts the sscanf calls-per-iteration down to two instead of five.

Also removed the old "objtxt" and "objmat" functions, it now uses the
proper name and supports the correct parameters.