Added sessions to avoid unnecessary obj. removals
Removed ZCMD dependency, no idea why it was still there.

Written a 'session' recording system. What this does is store a list of
removed objects for each player when they connect (or when the FS is
loaded for the first time). Then if the filterscript is reloaded at any
point, this data is used to avoid unnecessarily removing objects
multiple times. (Thanks to @Patrik356b for raising this issue)

The data is temporarily stored in binary format files named after
players for easy read/writing and it doesn't need to be human readable
anyway. A players session data file is removed when they disconnect.

Fixed a small issue with the debug modes and improved a few debug/info
prints.