Fixed a bug with material text parsing
Switched to CreateDynamicObject now since the draw distance parameter
was added to this function in the latest streamer plugin version. This
means the 1 cell large arrays can be removed for the world/interior
variables.

Also fixed some parsing errors with object material text, created a
separate parsing block for SetDynamicObjectMaterialText since the
parameter order is different.

Also foxed a minor issue with resolution names not loading correctly and
added support for raw integer input as the resolution parameter.